Fix mplib pointer math, replace an M2C_FIELD usage with a proper array access (#3442)
* Add sizeof usage and cleanup pointer math

* Fix array access instead of using M2C

* Remove M2C_FIELD usage from gmtoulib

* Remove M2C usages in crowdsfx

(cherry picked from commit 7e5738b0a8c7038c3cbd01497cfe713e8ad1b34e)

+++ src/melee/mp/mplib.c
@@ -1896,7 +1896,7 @@ bool mpCheckFloor(float ax, float ay, float bx, float by, float y_offset,
             float x1_sp40;
             float y1_sp3C;
             float dist2;
-            int line_offset;
+            ssize_t line_offset;
         block_8:
             if (cb != NULL && !cb(gobj, line_r26 - groundCollLine)) {
 #if BUILD_TARGET_PC
